requirements.txt
aiohttp==3.9.1
aiohttp-retry==2.8.3
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.2.0
async-timeout==4.0.3
attrs==23.2.0
azure-cognitiveservices-speech==1.34.0
certifi==2023.11.17
charset-normalizer==3.3.2
click==7.1.2
distro==1.9.0
exceptiongroup==1.2.0
Flask==1.1.4
Flask-Sockets==0.2.1
frozenlist==1.4.1
gevent==23.9.1
gevent-websocket==0.10.1
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.2
httpx==0.26.0
idna==3.6
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==2.0.1
multidict==6.0.4
numpy==1.26.3
openai==1.7.1
pydantic==2.5.3
pydantic_core==2.14.6
pydub==0.25.1
PyJWT==2.8.0
requests==2.31.0
sniffio==1.3.0
tqdm==4.66.1
twilio==8.11.0
typing_extensions==4.9.0
urllib3==2.1.0
Werkzeug==1.0.1
yarl==1.9.4
zope.event==5.0
zope.interface==6.1
